Adds include for stdint

+++ include/cat.h
@@ -2,6 +2,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <string.h>
 #include <unistd.h>
 
@@ -9,5 +10,5 @@
 #define MAX_BUFFER_SIZE 32
 
 static FILE *fopen_s(char *filename, char *open_mode);
-static void cat(FILE *in, int n);
+static void cat(FILE *in, uint8_t n);
 void usage(void);
